fix: rename templates.ts to templateHelpers.ts to avoid conflict with /templates folder
