+++ functions/src/emailNotifications/createModerationEmails.ts
@@ -4,7 +4,7 @@ import { IHowtoDB, IMapPin, IModerable } from '../../../src/models'
 import { IModerationStatus } from 'oa-shared'
 import { db } from '../Firebase/firestoreDB'
 import { DB_ENDPOINTS } from '../models'
-import * as templates from './templates'
+import * as templates from './templateHelpers'
 import { getUserAndEmail } from './utils'
 import { Change } from 'firebase-functions/v1'
 import { withErrorAlerting } from '../alerting/errorAlerting'
