+++ functions/src/emailNotifications/verifiedBadgeEmails.ts
@@ -2,7 +2,7 @@ import * as functions from 'firebase-functions'
 import { db } from '../Firebase/firestoreDB'
 import { getUserEmail } from './utils'
 import { DB_ENDPOINTS } from 'oa-shared'
-import { getUserVerifiedBadgeAddedEmail } from './templates'
+import { getUserVerifiedBadgeAddedEmail } from './templateHelpers'
 import type { IUserDB } from '../models'
 import { withErrorAlerting } from '../alerting/errorAlerting'
 import type { QueryDocumentSnapshot } from 'firebase-admin/firestore'
